requirements.txt
flask
pillow
requests
celery
